Added authentication with jwt, made it REST and removed unnecessary views
Added jwt authentication. Post put and delete are protected routes
REST
Removed views
Added CORS headers
Added error handling middleware

+++ controller/productController.js
@@ -1,5 +1,6 @@
 const Product = require("../models/product");
 const statusUtils = require("../utilities/statusUtils");
+const mongoose = require("mongoose");
 
 const { getProductNotFoundError, getServerError, PRODUCT_DELETED_MSG } =
   statusUtils;
@@ -10,6 +11,10 @@ const getAllProducts = async (_, res, next) => {
   } catch (err) {
     console.error(err);
     err = getServerError();
+    /**here since we are not inside any promise code, throw will also work intead of next(error)
+     *throw err
+     if we are throwing an error inside promise then catch,there it will not work as it is running asynchronously. There you have to use next(error)
+     */
     next(err);
   }
 };
@@ -17,6 +22,10 @@ const getAllProducts = async (_, res, next) => {
 const getProductById = async (req, res, next) => {
   try {
     const prdId = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(prdId)) {
+      const err = getProductNotFoundError();
+      throw err;
+    }
     const product = await Product.findById(prdId);
     if (!product) {
       const err = getProductNotFoundError();
@@ -41,7 +50,6 @@ const addProduct = async (req, res, next) => {
       imgUrl: req.body.imgUrl,
       description: req.body.description,
     });
-
     res.status(201).json(product);
   } catch (err) {
     console.error(err);
@@ -72,13 +80,17 @@ const updateProduct = async (req, res, next) => {
     if (!err.statusCode) {
       err = getServerError();
     }
-    next(err);
+    throw err;
   }
 };
 
 const deleteProduct = async (req, res, next) => {
   try {
     const prdId = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(prdId)) {
+      const err = getProductNotFoundError();
+      throw err;
+    }
     const product = await Product.findByIdAndDelete(prdId);
     if (!product) {
       const err = getProductNotFoundError();
